requirements.txt
# ── Demo parsing ──────────────────────────────────────────────────────────
demoparser2>=1.5.0
awpy>=2.0.0

# ── Core ML ───────────────────────────────────────────────────────────────
torch>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0

# ── Data storage ──────────────────────────────────────────────────────────
pyarrow>=15.0.0
fastparquet>=2024.2.0

# ── Config ────────────────────────────────────────────────────────────────
pyyaml>=6.0.1
omegaconf>=2.3.0

# ── Experiment tracking ───────────────────────────────────────────────────
tensorboard>=2.16.0

# ── Real-time / dashboard ─────────────────────────────────────────────────
flask>=3.0.0
streamlit>=1.33.0
onnxruntime>=1.17.0

# ── Dev / notebooks ───────────────────────────────────────────────────────
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
ipykernel>=6.29.0
tqdm>=4.66.0

# ── Testing ───────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0

# ── Demo downloading ───────────────────────────────────────────────────────
curl_cffi>=0.7.0
cloudscraper>=1.2.71
beautifulsoup4>=4.12.0
lxml>=5.2.0
rarfile>=4.1
playwright>=1.44.0
